AssistantPython/Assistant/requirements.txt
requests>=2.31
python-dotenv>=1.0
google-agents>=0.1
